+++ src/DreamFoodDelivery/DreamFoodDelivery.Domain.Logic/DreamFoodDelivery.Domain.Logic.csproj
@@ -14,6 +14,7 @@
       <PackageReference Include="AutoMapper" Version="9.0.0" />
       <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="7.0.0" />
       <PackageReference Include="FluentValidation" Version="8.6.2" />
+      <PackageReference Include="MethodBoundaryAspect.Fody" Version="2.0.118" />
       <PackageReference Include="Microsoft.CodeAnalysis.FxCopAnalyzers" Version="2.9.8">
         <PrivateAssets>all</PrivateAssets>
         <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
