+++ src/DreamFoodDelivery/DreamFoodDelivery.Web/Controllers/Menu/DishController.cs
@@ -1,0 +1,113 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using DreamFoodDelivery.Common;
+using DreamFoodDelivery.Domain.DTO;
+using DreamFoodDelivery.Domain.Logic.InterfaceServices;
+using FluentValidation.AspNetCore;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Swashbuckle.AspNetCore.Annotations;
+
+namespace DreamFoodDelivery.Web.Controllers.Menu
+{
+    [Route("api/[controller]")]
+    [ApiController]
+    public class DishController : ControllerBase
+    {
+        private readonly IMenuService _menuService;
+        public DishController(IMenuService menuService)
+        {
+            _menuService = menuService;
+        }
+
+        /// <summary>
+        /// Add dish
+        /// </summary>
+        /// <param name="dish"></param>
+        /// <returns></returns>
+        [HttpPost, Route("")]
+        [SwaggerResponse(StatusCodes.Status200OK, "Dish added")]
+        [SwaggerResponse(StatusCodes.Status400BadRequest, "Ivalid dish data")]
+        [LoggerAttribute]
+        public async Task<IActionResult> Create([FromBody, CustomizeValidator]DishToAdd dish)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            var result = await _menuService.AddAsync(dish);
+            return result.IsError ? BadRequest(result.Message) : (IActionResult)Ok(result.Data);
+        }
+
+        /// <summary>
+        /// Update dish
+        /// </summary>
+        /// <param name="dish">dish</param>
+        /// <returns>Dishes</returns>
+        [HttpPut, Route("")]
+        [SwaggerResponse(StatusCodes.Status400BadRequest, "Invalid paramater format")]
+        [SwaggerResponse(StatusCodes.Status404NotFound, "Dish doesn't exists")]
+        [SwaggerResponse(StatusCodes.Status200OK, "Dish updated", typeof(DishView))]
+        [SwaggerResponse(StatusCodes.Status500InternalServerError, "Something wrong")]
+        [LoggerAttribute]
+        public async Task<IActionResult> Update([FromBody, CustomizeValidator]DishToUpdate dish)
+        {
+            if (dish is null || !ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            try
+            {
+                var result = await _menuService.UpdateAsync(dish);
+                return result.IsError ? BadRequest(result.Message) : (IActionResult)Ok(result.Data);
+            }
+            catch (InvalidOperationException ex)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
+            }
+        }
+
+        /// <summary>
+        /// Delete dish
+        /// </summary>
+        /// <param name="id">dish id</param>
+        /// <returns></returns>
+        [HttpDelete, Route("{id}")]
+        [SwaggerResponse(StatusCodes.Status400BadRequest, "Ivalid ID")]
+        [SwaggerResponse(StatusCodes.Status404NotFound, "Dish doesn't exists")]
+        [SwaggerResponse(StatusCodes.Status200OK, "Dish deleted")]
+        [SwaggerResponse(StatusCodes.Status500InternalServerError, "Something goes wrong")]
+        [LoggerAttribute]
+        public async Task<IActionResult> RemoveById(string id)
+        {
+            if (string.IsNullOrEmpty(id) || !Guid.TryParse(id, out var _))
+            {
+                return BadRequest();
+            }
+            try
+            {
+                var result = await _menuService.RemoveByIdAsync(id);
+                return result.IsError ? BadRequest(result.Message) : (IActionResult)Ok(result.IsSuccess);
+            }
+            catch (InvalidOperationException ex)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
+            }
+        }
+
+        /// <summary>
+        /// Delete dishes
+        /// </summary>
+        /// <returns></returns>
+        [HttpDelete, Route("")]
+        [SwaggerResponse(StatusCodes.Status200OK, "Dishes removed")]
+        [LoggerAttribute]
+        public async Task<IActionResult> RemoveAllAsync()
+        {
+            await _menuService.RemoveAllAsync();
+            return Ok();
+        }
+    }
+}
