+++ src/DreamFoodDelivery/DreamFoodDelivery.Web/Controllers/Menu/TagController.cs
@@ -10,6 +10,7 @@ using Swashbuckle.AspNetCore.Annotations;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using FluentValidation.AspNetCore;
+using DreamFoodDelivery.Common;
 //using NSwag.Annotations;
 
 namespace DreamFoodDelivery.Web.Controllers
@@ -33,6 +34,7 @@ namespace DreamFoodDelivery.Web.Controllers
         [HttpGet, Route("")]
         [SwaggerResponse(StatusCodes.Status404NotFound, "There are no tags in list")]
         [SwaggerResponse(StatusCodes.Status200OK, "Tags were found", typeof(IEnumerable<TagView>))]
+        [LoggerAttribute]
         public async Task<IActionResult> GetAll()
         {
             var result = await _tagService.GetAllAsync();
@@ -47,6 +49,7 @@ namespace DreamFoodDelivery.Web.Controllers
         [HttpPost, Route("")]
         [SwaggerResponse(StatusCodes.Status200OK, "Tag added")]
         [SwaggerResponse(StatusCodes.Status400BadRequest, "Ivalid tag data")]
+        [LoggerAttribute]
         public async Task<IActionResult> Create([FromBody, CustomizeValidator]TagToAdd tag)
         {
             if (!ModelState.IsValid)
@@ -67,6 +70,7 @@ namespace DreamFoodDelivery.Web.Controllers
         [SwaggerResponse(StatusCodes.Status404NotFound, "Tag doesn't exists")]
         [SwaggerResponse(StatusCodes.Status200OK, "Tag updated", typeof(TagView))]
         [SwaggerResponse(StatusCodes.Status500InternalServerError, "Something wrong")]
+        [LoggerAttribute]
         public async Task<IActionResult> Update([FromBody, CustomizeValidator]TagToUpdate tag)
         {
 
@@ -95,6 +99,7 @@ namespace DreamFoodDelivery.Web.Controllers
         [SwaggerResponse(StatusCodes.Status404NotFound, "Tag doesn't exists")]
         [SwaggerResponse(StatusCodes.Status200OK, "Tag deleted")]
         [SwaggerResponse(StatusCodes.Status500InternalServerError, "Something goes wrong")]
+        [LoggerAttribute]
         public async Task<IActionResult> RemoveById(string id)
         {
             if (!Guid.TryParse(id, out var _) /*|| _orderService.GetById(id) == null*/ /*|| _commentService.GetById(id).UserId != UserId*/)
@@ -118,6 +123,7 @@ namespace DreamFoodDelivery.Web.Controllers
         /// <returns></returns>
         [HttpDelete, Route("")]
         [SwaggerResponse(StatusCodes.Status200OK, "Tags removed")]
+        [LoggerAttribute]
         public async Task<IActionResult> RemoveAllAsync()
         {
             await _tagService.RemoveAllAsync();
