+++ src/DreamFoodDelivery/DreamFoodDelivery.Web/Controllers/User/OrderController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using DreamFoodDelivery.Common;
 using DreamFoodDelivery.Domain.DTO;
 using DreamFoodDelivery.Domain.Logic.InterfaceServices;
 using FluentValidation.AspNetCore;
@@ -33,6 +34,7 @@ namespace DreamFoodDelivery.Web.Controllers
         [SwaggerResponse(StatusCodes.Status404NotFound, "There are no orders in list")]
         [SwaggerResponse(StatusCodes.Status200OK, "Orders were found", typeof(IEnumerable<OrderView>))]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [LoggerAttribute]
         public async Task<IActionResult> GetAll()
         {
             try
@@ -56,6 +58,7 @@ namespace DreamFoodDelivery.Web.Controllers
         [SwaggerResponse(StatusCodes.Status404NotFound, "order doesn't exists")]
         [SwaggerResponse(StatusCodes.Status200OK, "order was found", typeof(OrderView))]
         [SwaggerResponse(StatusCodes.Status500InternalServerError, "Something goes wrong")]
+        [LoggerAttribute]
         public async Task<IActionResult> GetById(string id)
         {
             if (string.IsNullOrEmpty(id) || !Guid.TryParse(id, out var _))
@@ -84,6 +87,7 @@ namespace DreamFoodDelivery.Web.Controllers
         [SwaggerResponse(StatusCodes.Status404NotFound, "User doesn't exists")]
         [SwaggerResponse(StatusCodes.Status200OK, "User updated")]
         [SwaggerResponse(StatusCodes.Status500InternalServerError, "Something wrong")]
+        [LoggerAttribute]
         public async Task<IActionResult> UpdateStatus([FromBody, CustomizeValidator]OrderToStatusUpdate orderStatus)
         {
             if (orderStatus is null || !ModelState.IsValid)
@@ -110,6 +114,7 @@ namespace DreamFoodDelivery.Web.Controllers
         [SwaggerResponse(StatusCodes.Status200OK, "order added")]
         [SwaggerResponse(StatusCodes.Status400BadRequest, "Ivalid order data")]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [LoggerAttribute]
         public async Task<IActionResult> Create([FromBody, CustomizeValidator]OrderToAdd order)
         {
             if (!ModelState.IsValid)
@@ -137,6 +142,7 @@ namespace DreamFoodDelivery.Web.Controllers
         [SwaggerResponse(StatusCodes.Status404NotFound, "Order doesn't exists")]
         [SwaggerResponse(StatusCodes.Status200OK, "Order updated", typeof(OrderToUpdate))]
         [SwaggerResponse(StatusCodes.Status500InternalServerError, "Something wrong")]
+        [LoggerAttribute]
         public async Task<IActionResult> Update([FromBody, CustomizeValidator]OrderToUpdate order)
         {
             if (order is null || !ModelState.IsValid)
@@ -163,6 +169,7 @@ namespace DreamFoodDelivery.Web.Controllers
         [SwaggerResponse(StatusCodes.Status404NotFound, "orders wasn't found")]
         [SwaggerResponse(StatusCodes.Status200OK, "ID users orders were found", typeof(IEnumerable<OrderView>))]
         [SwaggerResponse(StatusCodes.Status500InternalServerError, "Something goes wrong")]
+        [LoggerAttribute]
         public async Task<IActionResult> GetByUserId(string id)
         {
             //if (User.Identity.IsAuthenticated)
@@ -194,6 +201,7 @@ namespace DreamFoodDelivery.Web.Controllers
         [SwaggerResponse(StatusCodes.Status404NotFound, "order doesn't exists")]
         [SwaggerResponse(StatusCodes.Status200OK, "order deleted")]
         [SwaggerResponse(StatusCodes.Status500InternalServerError, "Something goes wrong")]
+        [LoggerAttribute]
         public async Task<IActionResult> RemoveById(string id)
         {
             if (!string.IsNullOrEmpty(id) || !Guid.TryParse(id, out var _))
