+++ src/DreamFoodDelivery/DreamFoodDelivery.Web/FodyWeavers.xml
@@ -1,0 +1,3 @@
+﻿<Weavers xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="FodyWeavers.xsd">
+  <MethodBoundaryAspect />
+</Weavers>
